fix: scrollBehavior on BookDetail LargeTopAppBar, snackbar duration, MemoryCard icon scale animation
- Add exitUntilCollapsedScrollBehavior() to BookDetailScreen LargeTopAppBar so the collapsing header actually responds to scroll
- Shorten snackbar duration from Long to Short in BookDetailScreen and BooksScreen (4.5s was too long for non-critical messages)
- Move graphicsLayer scale animation from IconButton to Icon in MemoryCard edit/share/delete actions (was scaling entire 36dp touch target instead of 18dp icon)
- Add shimmer loop count 1→3 for more visible/lively loading animation on HomeScreen skeletons

+++ app/src/main/kotlin/com/memoryproject/app/ui/books/BooksScreen.kt
@@ -67,7 +67,7 @@ fun BooksScreen(
     // Show error as snackbar
     LaunchedEffect(uiState.error) {
         uiState.error?.let {
-            snackbarHostState.showSnackbar(it, duration = SnackbarDuration.Long)
+            snackbarHostState.showSnackbar(it, duration = SnackbarDuration.Short)
             viewModel.clearError()
         }
     }
